+++ examples/wrap_map.rs
@@ -46,7 +46,7 @@ fn setup_grid(
     mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
     let layout = HexLayout {
-        hex_size: HEX_SIZE,
+        scale: HEX_SIZE,
         ..default()
     };
     let mesh = meshes.add(hexagonal_plane(&layout));